requirements.txt
####### required packages #######
pytest
GitPython